+++ host-configs/lc/toss_4_x86_64_ib/clang.cmake
@@ -1,0 +1,20 @@
+##############################################################################
+# Copyright (c) 2016-25, Lawrence Livermore National Security, LLC and CHAI
+# project contributors. See the CHAI LICENSE file for details.
+#
+# SPDX-License-Identifier: BSD-3-Clause
+##############################################################################
+
+# Set up software versions
+set(CLANG_VERSION "14.0.6" CACHE PATH "")
+set(GCC_VERSION "12.1.1" CACHE PATH "")
+
+# Set up compilers
+set(COMPILER_BASE "/usr/tce/packages/clang/clang-${CLANG_VERSION}-magic" CACHE PATH "")
+set(CMAKE_C_COMPILER "${COMPILER_BASE}/bin/clang" CACHE PATH "")
+set(CMAKE_CXX_COMPILER "${COMPILER_BASE}/bin/clang++" CACHE PATH "")
+
+# Set up compiler flags
+set(GCC_HOME "/usr/tce/packages/gcc/gcc-${GCC_VERSION}-magic" CACHE PATH "")
+set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} --gcc-toolchain=${GCC_HOME}" CACHE STRING "")
+set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} --gcc-toolchain=${GCC_HOME}" CACHE STRING "")