+++ RELEASE_NOTES.md
@@ -13,6 +13,25 @@ in this file.
 
 The format of this file is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [Version 2025.03.0] - Release date 2025-03-19
+
+### Added
+- Added a ManagedArray::clone function and deprecated chai::deepCopy.
+
+### Fixed
+- Fixes reallocate when using pinned or unified memory.
+- Adds missing synchronize when using pinned memory.
+- Fixes possible hangs when evicting data.
+- Now respects allocators passed to ManagedArray constructors when CHAI\_DISABLE\_RM=TRUE.
+
+### Removed
+- Removes deprecated ManagedArray::getPointer method. Use ManagedArray::data instead.
+- Removes ManagedArray::incr and ManagedArray::decr methods. Use ManagedArray::pick and ManagedArray::set instead.
+- Removes optional support for implicitly casting between raw pointers and ManagedArrays (CHAI\_ENABLE\_IMPLICIT\_CONVERSIONS). Use makeManagedArray and ManagedArray::data to perform explicit conversions instead.
+- Removes equality and inequality comparison operators between ManagedArrays and raw pointers.
+- Removes make\_managed\_from\_factory function for creating managed\_ptr objects from factory functions. This change will lead to safer adoption of allocators during construction and destruction of managed\_ptr objects.
+- Removes CHAI\_ENABLE\_PICK CMake option. ManagedArray::pick and ManagedArray::set will always be available.
+
 ## [Version 2024.07.0] - Release date 2024-07-26
 
 ### Changed
@@ -35,7 +54,7 @@ The format of this file is based on [Keep a Changelog](http://keepachangelog.com
 ## [Version 2024.02.0] - Release date 2024-03-04
 
 ### Added
-- Support for APUs with a single memory space. To use, configure with -DCHAI\_DISABLE\_RM=ON -DCHAI\_GPU\_THIN\_ALLOCATE=ON.
+- Support for APUs with a single memory space. To use, configure with -DCHAI\_DISABLE\_RM=ON -DCHAI\_THIN\_GPU\_ALLOCATE=ON.
 
 ### Changed
 - Moved installed CMake targets from share/chai/cmake to lib/cmake/chai to be consistent with other libraries in the RAJA Portability Suite